edit llama

+++ models/llama-3.2-1B.py
@@ -3,7 +3,7 @@ import torch, transformers
 from ._hf_wrapper import HFWrapper
 ID, _SEQ = "meta-llama/Llama-3.2-1B", 32
 
-def get_model(): return HFWrapper(transformers.AutoModel.from_pretrained(ID).eval())
+def get_model(): return HFWrapper(transformers.AutoModel.from_pretrained(ID, torch_dtype=torch.bfloat16).eval())
 def get_dummy_input():
     tok = transformers.AutoTokenizer.from_pretrained(ID)
     vocab = tok.vocab_size
